Crate tagua_llvm

Tagua VM

Tagua VM is an experimental PHP Virtual Machine written with the Rust language and the LLVM Compiler Infrastructure.

This library defines safe bindings to LLVM designed for Tagua VM. The API does not aim at being a generic safe bindings implementation. One might not feel lost if already familiar with LLVM infrastructure. The basis of the infrastructure is the context and the module, functions are declared inside a module and the execution engine is MCJIT (see MCJIT Design and Implementation).

Modules

builder
Build blocks, statements, … of all kinds.

context
Context manipulation.

engine
Execute the VM intermediate representation.

function
Function manipulation.

module
Module manipulation.

native_type
LLVM types manipulation and bindings from Rust types.

value
Value manipulation.

Traits

LLVMRef